refactor: remove stale files
refactor: add types folder
