+++ types/question.ts
@@ -1,0 +1,4 @@
+export enum QuestionStep {
+  AnswerQuestion = 1,
+  PickPercentage = 2,
+}
